Rename utils wildcard hooks
