+++ client/wildcard/src/hooks/index.ts
@@ -4,5 +4,6 @@ export { useOffsetPagination } from './useOffsetPagination'
 export { useSearchParameters } from './useSearchParameters'
 export { useAutoFocus } from './useAutoFocus'
 export { useInterval } from './useInterval'
-export { useKeyboard } from './use-keyboard'
-export { useOnClickOutside } from './use-outside-click'
+export { useKeyboard } from './useKeyboard'
+export { useOnClickOutside } from './useOnClickOutside'
+export { useDeepMemo } from './useDeepMemo'
