+++ client/wildcard/src/hooks/useOnClickOutside.ts
